// src/utils/types/CharacterType.ts
export interface Hero {
    name: string
    height: string
    mass: string
    hair_color: string
    skin_color: string
    eye_color: string
    birth_year: string
    gender: string
    homeworld: string
    films: string[]
    species: any[]
    vehicles: any[]
    starships: any[]
    created: string
    edited: string
    url: string
}